Enqueue an empty job to ensure the background job worker is alive
