+++ modules/EasyAbp.PaymentService.WeChatPay/src/EasyAbp.PaymentService.WeChatPay.Domain/EasyAbp/PaymentService/WeChatPay/Background/EmptyJobArgs.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using Volo.Abp.MultiTenancy;
+
+namespace EasyAbp.PaymentService.WeChatPay.Background;
+
+[Serializable]
+public class EmptyJobArgs : IMultiTenant
+{
+    public Guid? TenantId { get; set; }
+
+    public EmptyJobArgs(Guid? tenantId)
+    {
+        TenantId = tenantId;
+    }
+}
